requirements.txt
@"
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
Werkzeug==3.0.3
tensorflow==2.16.1
numpy>=1.26.4
Pillow==10.3.0
waitress==2.1.2
pyopenssl==24.1.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
tzdata
requests==2.31.0
huggingface_hub>=0.20.0
"@ | Out-File -FilePath requirements.txt -Encoding utf8
